Split out braiding mod and add some comments
